NFmiPressTimeDescription::PreProcessDefinition -metodissa muutettu unsigned long pos -> size_t pos
, jotta std::string find:in palauttama arvo toimii myös 64-bit versiossa. Winkkarissa (unsigned) long on 32-bit ja linuxissa 64-bit.

+++ source/NFmiPressTimeDescription.cpp
@@ -409,7 +409,7 @@ bool NFmiPressTimeDescription::PreProcessDefinition(const string & inFileName,
 	file.close();
 	file.clear();
 
-	  unsigned long pos = outString.find("$#");
+	  size_t pos = outString.find("$#");
 	  if (pos != string::npos)
 	  {
 		  string newString = outString.substr(pos, 3);
